Combat system 0.1 is ready
Okay i have a very basic combat system laid out, it is not ready to be tested in game yet, i will be doing the exploring & looting world state now and then once complete we should have all the peices we need to actually create a little test game. If that succeeds then its o ff to the races of  connecting the story to indidual back end scenes/triggers/enviroments.

The more i code the more I realize how much there is to even consider for a text based RPG so if i missed something or you see an error in my code please point it out. "

Also just like the item values, these combat values (damage, guard, etc) are NOT balanced in the slightest and MUST BE FIXED before we put the game into full prod because right now I have formulas set up exactly how I want them but the numbers in them are random.