should be ll the files we need
These are no where near done yet but i got a burst of inspiration as I started the exploration classes and thought of all the missing functionality that would be necessary for the game to run. I believe everything needed is here, at least its bones that is.

Then the next move once these are all finished is to actually connect them to the story, and then the gui. If anything is missing thats not here please add it.

The .txt files have placeholders only, i made it so both scenes (description of where a player is at) and dialogue (speaking with npcs) is living in a .txt with hashmaps. This should in theory make adding the actual story scenes together much easier. We will just assign unique ids to each and then have them trigger based on whatever check we put in. One think we will need to do is like tie different player choices (in terms of actions on how they want to proceed, looting, dialogue, going to new area, etc) to options, i was thinking we could do this by assigning like a  combat id, looting id, speaking id, and then exploring id and then essentially like those to instances of each so that when they are selected the correct world state loads in. Im not sure if that makes sense but in my head it does so if theres any questions please ask.