Should not have waited multiple days to commit, my bad
Changed out ExploringController for SceneManager, fixed some formatting stuff and finally made it so the scenes can be connected and navigated between by the user. Made some other small polished to the code but now i beleive (minus the ability for players  to equipt items) literally all the peices are here.

I wrote a comprehensive readMe to completely describe the code in a more digestible way as well as gave instructions  on how to go about adding each of the necessary story elements for the game. Now the entire game should be able to be put together story wise from these txts, and then as long as the instructions were followed closely then itll just be a matter of connecting the gui and making some small adjustments based off specific events the story  team wants to add.

Once the story team is finished with everything then the only thing i can think of is a final controller that helps switch world state enums back and forth