It's Alive!
We have done it, the game including all three game states function completely now... you can test the game yourself even! Just compile/run the sceneTester.java

Is the game balanced? no

Is it missing scenes? Sure

BUT it runs! So thats cool. I basically had to edit and adjust almost  every class, or well it felt that way. So basically a lot got changed and adjusted... a lot but it works now so thats what matters.